# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BSV blockchain indexer API for transaction outputs, blocks, and related data",
        "title": "1Sat Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/shruggr/1sat-indexer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "5.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v5/spends": {
            "post": {
                "description": "Get spend information for multiple transaction outputs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spends"
                ],
                "summary": "Get multiple spend records",
                "parameters": [
                    {
                        "description": "Array of outpoints",
                        "name": "outpoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spend information array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v5/tx": {
            "post": {
                "description": "Broadcast a transaction to the BSV network",
                "consumes": [
                    "application/octet-stream",
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Broadcast transaction",
                "parameters": [
                    {
                        "enum": [
                            "beef"
                        ],
                        "type": "string",
                        "description": "Transaction format: 'beef' or standard",
                        "name": "fmt",
                        "in": "query"
                    },
                    {
                        "description": "Transaction bytes (binary or hex)",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Broadcast response",
                        "schema": {
                            "$ref": "#/definitions/broadcast.BroadcastResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid transaction",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v5/txo": {
            "post": {
                "description": "Get multiple transaction outputs by outpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "txos"
                ],
                "summary": "Get multiple transaction outputs",
                "parameters": [
                    {
                        "description": "Array of outpoints",
                        "name": "outpoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include spend information",
                        "name": "spend",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "broadcast.BroadcastResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "txid": {
                    "type": "string"
                }
            }
        },
        "evt.Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "idx.IndexData": {
            "type": "object",
            "properties": {
                "data": {},
                "deps": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/evt.Event"
                    }
                }
            }
        },
        "idx.Txo": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/idx.IndexData"
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "idx": {
                    "type": "integer"
                },
                "outpoint": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "owners": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "satoshis": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "script": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "spend": {
                    "type": "string"
                }
            }
        }
    }
}